backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
